Move process_frame to the binaries
Make the library code not depend to the y4m crate.
