+++ src/bin/rav1repl.rs
@@ -11,6 +11,10 @@ extern crate rustyline;
 extern crate y4m;
 
 extern crate rav1e;
+
+mod common;
+use common::*;
+
 use rav1e::*;
 
 use rustyline::error::ReadlineError;
